mythmusic: Save playlists to DB when changed not at frontend exit
This changes things so any changes to playlists are saved to the database
immediately rather than cached and saved when the frontend exits.

This fixes a problem were you could lose edits to a playlist if the FE is not
shutdown cleanly like when using the idle timer to shutdown or if you shutdown
the computer from the FE exit menu.