MythBackend: add a CHANGE_STORAGE_GROUP command to the myth protocol
This command allows one or more recordings to be moved from there current
storage group to the new one specified in the command. This also moves any
preview images generated for the recording.

The command syntax is:-

CHANGE_STORAGE_GROUP newgroupname chanid1 starttime1 chanid2 starttime2 ...

Where :-

newgroupname - is the name of the storage group to move the recordings to.
chanid1      - is the chanid of the recording
starttime1   - is the start time of the recording in # of seconds since
               Jan 1, 1970, in UTC time.

Repeat chanid and startime pairs to move additional recordings.