Watch Recordings: add option to move recordings to a different storage group
This adds a "Change Storage Group" menu option to move individual or several
recordings in a playlist to a different storage group.

My use case is to move recordings from the 'default' storage group to my
"Offline_Storage" storage group which is a removable usb drive in a cradle
allowing me to quickly add extra storage.

NOTE: currently it only supports moving recordings on the master BE although
you can use a remote FE to chose which recordings to move.