mythutil: add an option to check recordings
This adds an option to check the integrity of all recordings. Currently it
checks if the file exists, they are not zero bytes in size and they have a
seek table. You can optionally fix the seek tables of any recordings were it
is missing.

You run the command like this:

mythutil -v none --checkrecordings --fixseektable