mythfrontend_qml: add MythIncrementalModel base class
This provides most of the functionality to download from a master backend
using the Services API endpoints that support the paged downloads using
StartIndex, Count and TotalAvailable properties so only the required data
is actually fetched.

Derived classes have to implement the two abstract functions StartDownload()
and ProcessDownload() that setup the URL to download from and process the
downloaded XML extracting the data required by the model.