settings: new setting to set how long to wait before shutting down a frontend
This adds a new setting :-
Settings -> Shutdown/Reboot Settings -> Time to wait before attempting to
                                        shutdown the frontend system (seconds)

This sets the number of seconds to wait when the frontend is in standby mode
before attempting to shutdown the system (only applicable if this is a
frontend only system). A value of zero prevents the frontend automatically
shutting down the system.