MythZoneminder: add support for ZoneMinder 1.32.x
This updates mythzmserver to work with the current stable version of
ZoneMinder (tested on 1.32.3).

NOTES:
By default mythzmserver will read ZM's config from /etc/zm/zm.conf and
/etc/zm/conf.d/01-system-paths.conf which now stores more information
like the location of the mmap files and image storage locations. You
can override the location of these config files using -c/--zmconfig and
-o/--zmoverrideconfig command line options to mythzmserver.

Previous versions of mythzmserver used the config setting ZM_VERSION to
determine what version of ZM was running. Later versions of ZM have removed
it from the config files and in fact if it's present it breaks ZM's API so
should be removed. For now if not present we assume we are dealing with
ZM >= 1.32.x